Structural and Styling changes
- Adds folder for Landing page
- Adds LanginPage.styling.tsx
- Refactors Landingpage to use Container component
