+++ frontend/src/pages/index.tsx
@@ -1,6 +1,6 @@
 import Houses from './Houses/Houses';
 import Login from './Login';
-import LandingPage from './Landingpage';
+import LandingPage from './LandingPage/Landingpage';
 import PostRegister from './PostRegister';
 
 export { Houses, Login, LandingPage, PostRegister };
